+++ .github/workflows/build.yml
@@ -5,6 +5,9 @@ on: [push, pull_request]
 jobs:
   build:
     runs-on: windows-latest
+    defaults:
+      run:
+        working-directory: open-cue-service
     steps:
       - uses: actions/checkout@v2
       - name: Setup .NET Core
