+++ .github/workflows/pages.yml
@@ -8,6 +8,9 @@ on:
 jobs:
   generate-openapi-spec:
     runs-on: windows-latest
+    defaults:
+      run:
+        working-directory: open-cue-service
     steps:
       - uses: actions/checkout@v2
       - name: Setup .NET Core
@@ -28,6 +31,9 @@ jobs:
   deploy:
     needs: generate-openapi-spec
     runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: open-cue-service
     steps:
       - name: Download openapi.json
         uses: actions/download-artifact@v2
