+++ README.md
@@ -51,5 +51,5 @@ Restart iCUE and Open CUE Service after you have changed any file.
 ## License
 This project is distributed under [Apache License, Version 2.0](LICENSE).
 
-The software iCUE and the file CgSDK.x64_2015.dll are supplied by Corsair Components, Inc. and distributed under different terms.
+The software iCUE and the file `lib/CgSDK.x64_2015.dll` are supplied by Corsair Components, Inc. and distributed under different terms.
 For more information, contact Corsair Components, Inc.
