+++ open-cue-sdk/src/CorsairGameSdk.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace OpenCueService.CgSdk
+namespace OpenCue.Sdk
 {
     public class CorsairGameSdk
     {
