+++ open-cue-service/src/Config.cs
@@ -1,4 +1,4 @@
-namespace OpenCueService
+namespace OpenCue.Service
 {
     public class Config
     {
