+++ open-cue-service/src/Controllers/ApiErrorFilter.cs
@@ -1,10 +1,9 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 
-namespace OpenCueService.Controllers
+namespace OpenCue.Service.Controllers
 {
-    using CgSdk;
-    public class SdkExceptionFilter : IActionFilter, IOrderedFilter
+    public class ApiErrorFilter : IActionFilter, IOrderedFilter
     {
         public int Order { get; set; } = int.MaxValue - 10;
 
@@ -12,11 +11,11 @@ namespace OpenCueService.Controllers
 
         public void OnActionExecuted(ActionExecutedContext context)
         {
-            if (context.Exception is SdkError exception)
+            if (context.Exception is ApiError apiError)
             {
-                context.Result = new ObjectResult(exception.Message)
+                context.Result = new ObjectResult(apiError.HttpMessage)
                 {
-                    StatusCode = 400,
+                    StatusCode = apiError.StatusCode,
                 };
                 context.ExceptionHandled = true;
             }
