+++ open-cue-service/src/Startup.cs
@@ -4,7 +4,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace OpenCueService
+namespace OpenCue.Service
 {
     using Controllers;
     public class Startup
